add infer_cfg export in slim export_model (#1554)
